+++ webpack.config.js
@@ -1,27 +1,29 @@
 const path = require("path");
+const nodeExternals = require("webpack-node-externals");
 
 module.exports = {
   entry: "./src/index",
   output: {
     filename: "fs-pro.min.js",
     path: path.join(__dirname, "dist"),
-    libraryTarget: "commonjs"
+    libraryTarget: "commonjs",
   },
+  externals: [nodeExternals()],
   target: "node",
   node: {
     fs: "empty",
-    path: "empty"
+    path: "empty",
   },
   mode: "production",
   module: {
     rules: [
       {
         test: /\.ts$/,
-        use: "ts-loader"
-      }
-    ]
+        use: "ts-loader",
+      },
+    ],
   },
   resolve: {
-    extensions: [".ts", ".js"]
-  }
+    extensions: [".ts", ".js"],
+  },
 };